+++ src/core/ObservableQuery.ts
@@ -125,7 +125,7 @@ export class ObservableQuery<
     });
   }
 
-  public getCurrentResult(): ApolloQueryResult<TData> {
+  public getCurrentResult(saveAsLastResult = true): ApolloQueryResult<TData> {
     const { lastResult, lastError } = this;
     const networkStatus = this.queryInfo.networkStatus || NetworkStatus.ready;
     const result: ApolloQueryResult<TData> = {
@@ -141,7 +141,9 @@ export class ObservableQuery<
     const { fetchPolicy = 'cache-first' } = this.options;
     if (fetchPolicy === 'no-cache' ||
         fetchPolicy === 'network-only') {
-      result.partial = false;
+      // Similar to setting result.partial to false, but taking advantage
+      // of the falsiness of missing fields.
+      delete result.partial;
     } else if (
       !result.data ||
       // If this.options.query has @client(always: true) fields, we cannot
@@ -154,24 +156,29 @@ export class ObservableQuery<
       !this.queryManager.transform(this.options.query).hasForcedResolvers
     ) {
       const diff = this.queryInfo.getDiff();
-      result.partial = !diff.complete;
       result.data = (
         diff.complete ||
         this.options.returnPartialData
       ) ? diff.result : void 0;
-      // If the cache diff is complete, and we're using a FetchPolicy that
-      // terminates after a complete cache read, we can assume the next
-      // result we receive will have NetworkStatus.ready and !loading.
-      if (diff.complete &&
-          result.networkStatus === NetworkStatus.loading &&
-          (fetchPolicy === 'cache-first' ||
-           fetchPolicy === 'cache-only')) {
-        result.networkStatus = NetworkStatus.ready;
-        result.loading = false;
+      if (diff.complete) {
+        // If the diff is complete, and we're using a FetchPolicy that
+        // terminates after a complete cache read, we can assume the next
+        // result we receive will have NetworkStatus.ready and !loading.
+        if (result.networkStatus === NetworkStatus.loading &&
+            (fetchPolicy === 'cache-first' ||
+             fetchPolicy === 'cache-only')) {
+          result.networkStatus = NetworkStatus.ready;
+          result.loading = false;
+        }
+        delete result.partial;
+      } else {
+        result.partial = true;
       }
     }
 
-    this.updateLastResult(result);
+    if (saveAsLastResult) {
+      this.updateLastResult(result);
+    }
 
     return result;
   }
@@ -271,36 +278,11 @@ export class ObservableQuery<
 
     const qid = this.queryManager.generateQueryId();
 
+    // Simulate a loading result for the original query with
+    // result.networkStatus === NetworkStatus.fetchMore.
     if (combinedOptions.notifyOnNetworkStatusChange) {
-      const currentResult = this.getCurrentResult();
-
-      // If we neglect to update queryInfo.networkStatus here,
-      // getCurrentResult may return a loading:false result while
-      // fetchMore is in progress, since getCurrentResult also consults
-      // queryInfo.networkStatus. Note: setting queryInfo.networkStatus
-      // to an in-flight status means that QueryInfo#shouldNotify will
-      // return false while fetchMore is in progress, which is why we
-      // call this.reobserve() explicitly in the .finally callback after
-      // fetchMore (below), since the cache write will not automatically
-      // trigger a notification, even though it does trigger a cache
-      // broadcast. This is a good thing, because it means we won't see
-      // intervening query notifications while fetchMore is pending.
       this.queryInfo.networkStatus = NetworkStatus.fetchMore;
-
-      // Simulate a loading result for the original query with
-      // networkStatus === NetworkStatus.fetchMore.
-      this.observer.next!({
-        // Note that currentResult is an ApolloCurrentQueryResult<TData>,
-        // whereas this.observer.next expects an ApolloQueryResult<TData>.
-        // Fortunately, ApolloCurrentQueryResult is a subtype of
-        // ApolloQueryResult (with additional .error and .partial fields),
-        // so TypeScript has no problem with this sleight of hand.
-        // TODO Consolidate these two types into a single type (most
-        // likely just ApolloQueryResult) after AC3 is released.
-        ...currentResult,
-        loading: true,
-        networkStatus: NetworkStatus.fetchMore,
-      });
+      this.observe();
     }
 
     return this.queryManager.fetchQuery(
@@ -601,8 +583,18 @@ once, rather than every time you call fetchMore.`);
     return this.getReobserver().reobserve(newOptions, newNetworkStatus);
   }
 
-  private observer: Observer<ApolloQueryResult<TData>> = {
-    next: result => {
+  // Pass the current result to this.observer.next without applying any
+  // fetch policies, bypassing the Reobserver.
+  private observe() {
+    // Passing false is important so that this.getCurrentResult doesn't
+    // save the fetchMore result as this.lastResult, causing it to be
+    // ignored due to the this.isDifferentFromLastResult check in
+    // this.observer.next.
+    this.observer.next(this.getCurrentResult(false));
+  }
+
+  private observer = {
+    next: (result: ApolloQueryResult<TData>) => {
       if (this.lastError || this.isDifferentFromLastResult(result)) {
         this.updateLastResult(result);
         iterateObserversSafely(this.observers, 'next', result);