+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Disable feud-stopping logic after any cache eviction. <br/>
   [@benjamn](https://github.com/benjamn) in [#6817](https://github.com/apollographql/apollo-client/pull/6817)
 
+- Prevent full reobservation of queries affected by optimistic mutation updates, while still delivering results from the cache. <br/>
+  [@benjamn](https://github.com/benjamn) in [#6854](https://github.com/apollographql/apollo-client/pull/6854)
+
 ## Apollo Client 3.1.3
 
 ## Bug Fixes